+++ message_history/src/main/java/ru/vyapps/vkgram/message_history/Message.kt
@@ -3,9 +3,9 @@ package ru.vyapps.vkgram.message_history
 import java.util.*
 
 data class Message(
-    val id: Long,
-    val userId: Long,
-    val conversationId: Long,
+    val id: Int,
+    val userId: Int,
+    val conversationId: Int,
     val text: String,
     val date: Date,
     val out: Int
